% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hcdp_rainfall_st.R
\name{get_hcdp_rainfall}
\alias{get_hcdp_rainfall}
\title{Download HCDP Rainfall Data (Station)}
\usage{
get_hcdp_rainfall(
  year,
  month = NULL,
  fill = "partial",
  period = "month",
  extent = "statewide",
  outdir = ""
)
}
\arguments{
\item{year}{Year of the data (e.g., 2020).}

\item{month}{Month of the data (e.g., "03" for March) # if provided, download daily data for the month}

\item{fill}{Optional. Used for DAILY station data, such as "partial" or "raw". Defaults to "partial". raw: Unfilled data that has not undergone QA/QC. partial: Partially filled data that has undergone QA/QC and had statistical methods applied to it to estimate some missing values}

\item{period}{The period of data (default: "month"). Use "day" for daily data.}

\item{extent}{The geographical extent of the data (default: "statewide").}
}
\value{
The file name of the downloaded file.
}
\description{
This function downloads HCDP rainfall data for a specified year, month, and optional day.
}
\examples{
# Download daily rainfall for March 2020 to the outdir
get_hcdp_rainfall(year = 2020, month = 3, outdir="~/Downloads")
# Download monthly rainfall for 2020 to the outdir
get_hcdp_rainfall(year = 2020, outdir="~/Downloads")

}
