% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_raster.R
\name{get_timeseries_raster}
\alias{get_timeseries_raster}
\title{Download Raster Timeseries Data from the HCDP API}
\usage{
get_timeseries_raster(
  start,
  end,
  datatype,
  extent,
  row = NULL,
  col = NULL,
  index = NULL,
  lat = NULL,
  lon = NULL,
  production = NULL,
  period = NULL,
  token
)
}
\arguments{
\item{start}{Start date for the timeseries (format: "YYYY-MM").}

\item{end}{End date for the timeseries (format: "YYYY-MM").}

\item{datatype}{The type of data to retrieve (e.g., "rainfall" or "temperature").}

\item{extent}{The spatial extent of the raster (e.g., "bi", "statewide").}

\item{row}{Optional. The row index for the grid cell to produce a timeseries.}

\item{col}{Optional. The column index for the grid cell to produce a timeseries.}

\item{index}{Optional. The 1D index of the value to produce a timeseries (width * row + col).}

\item{lat}{Optional. Latitude of the location to produce a timeseries.}

\item{lon}{Optional. Longitude of the location to produce a timeseries.}

\item{production}{Optional. Specifies data production type if required (e.g., "new").}

\item{period}{Optional. Specifies the period of data (e.g., "month", "day").}

\item{token}{Your personal access token for authentication.}
}
\value{
A data frame containing datetime-value pairs for the specified timeseries.
}
\description{
This function queries the HCDP API to retrieve a timeseries of raster data (e.g., rainfall or temperature)
for a specified location and date range.
}
\examples{
\dontrun{
  get_timeseries_raster(
    start = "2020-01",
    end = "2023-06",
    datatype = "rainfall",
    extent = "statewide",
    lat = 21.539576,
    lon = -157.965820,
    production = "new",
    period = "month",
    token = "YOUR_ACCESS_TOKEN"
  )
}
}
